+++ includes/class-predis.php
@@ -53,6 +53,7 @@ class Predis {
             'port',
             'path',
             'password',
+            'sentinel_password',
             'database',
             'timeout',
             'read_timeout',
@@ -68,6 +69,9 @@ class Predis {
 
         if ( isset( $parameters['password'] ) && $parameters['password'] === '' ) {
             unset( $parameters['password'] );
+        }
+        if ( isset( $parameters['sentinel_password'] ) && $parameters['sentinel_password'] === '' ) {
+            unset( $parameters['sentinel_password'] );
         }
 
         if ( defined( 'WP_REDIS_SHARDS' ) ) {
@@ -108,6 +112,37 @@ class Predis {
                         $options['parameters']['password'] = WP_REDIS_PASSWORD[1];
                     } else {
                         $options['parameters']['password'] = WP_REDIS_PASSWORD;
+                    }
+                }
+
+                // Add support for password in Sentinel parameters when using Predis
+                if ( $constant == 'WP_REDIS_SERVERS' && defined( 'WP_REDIS_SENTINEL' ) && isset( $parameters['sentinel_password'] ) ) {
+                    foreach( $servers as $index => $server ) {
+                        // skip any connect string that already has a password, e.g. tcp://1.2.3.4:26379?password=abc123
+                        if ( strpos( $server, 'password=' ) !== false ) {
+                            continue;
+                        }
+                        // convert server string to array of components to add password
+                        $urlParts = parse_url( $server );
+                        if ( !$urlParts ) {
+                            continue; // can't parse this server string, skip it
+                        }
+                        $entry  = [
+                            'scheme' => $urlParts['scheme'] ?? 'tcp',
+                            'host'   => $urlParts['host'] ?? 'localhost',
+                            'port'   => $urlParts['port'] ?? 26379,
+                        ];
+                        if ( is_array( $parameters['sentinel_password'] ) ) {
+                            $entry['username'] = WP_REDIS_SENTINEL_PASSWORD[0];
+                            $entry['password'] = WP_REDIS_SENTINEL_PASSWORD[1];
+                        } else {
+                            $entry['password'] = WP_REDIS_SENTINEL_PASSWORD;
+                        }
+                        $servers[$index] = $entry; // replace server string with array of connection parameters
+                    }
+
+                    if ( ! empty( $parameters['timeout'] ) ) {
+                        $options['parameters']['timeout'] = $parameters['timeout'];
                     }
                 }
             }