+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 - Delete transients when cache is enabled
 - Respect `WP_REDIS_SSL_CONTEXT` when using Predis and clustered connections
 
+# 2.7.1
+
+- Support password configuration for Sentinel connections when using Predis (from @bsabalaskey)
+- Configuration uses `WP_REDIS_SENTINEL_PASSWORD` constant
+
 ## 2.7.0
 
 - Preserve key TTL when calling (in|de)crement methods